Moving to MetadataUpload
